Refactor validation imports and enhance user fetching in auth middleware
