+++ services/identity-service/middleware/auth.middleware.js
@@ -1,15 +1,24 @@
 import { verifyToken } from "../config/jwt.js";
+import { prisma } from "../config/database.js";
 
-export const authMiddleware = (req, res, next) => {
-  const token = req.headers.authorization?.split(" ")[1];
-  if (!token) return res.status(401).json({ error: "Unauthorized" });
-
+export const authMiddleware = async (req, res, next) => {
   try {
+    const token = req.headers.authorization?.split(" ")[1];
+    if (!token) return res.status(401).json({ error: "Unauthorized" });
+
     const decoded = verifyToken(token);
-    req.user = decoded;
+    
+    // Fetch the latest user data from the database
+    const user = await prisma.user.findUnique({ where: { id: decoded.id } });
+
+    if (!user) {
+      return res.status(401).json({ error: "User not found" });
+    }
+
+    req.user = user; // Attach the latest user data
     next();
-  // eslint-disable-next-line no-unused-vars
   } catch (err) {
+    console.error("Authentication error:", err);
     res.status(401).json({ error: "Invalid token" });
   }
 };
